requirements.txt
aiodns==1.1.1
aiohttp==3.4.4
black==18.9b0
cchardet==2.1.4
Click==7.0
mypy==0.641
pytest==3.9.3
PyYAML==3.13
flake8==3.6.0
ujson==1.35
cython==0.29
marshmallow==2.16.3
raven==6.9.0
raven-aiohttp==0.7.0